Move document format into Card subclasses
