+++ src/main/java/editor/database/card/MultiCard.java
@@ -12,6 +12,9 @@ import java.util.TreeMap;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import javax.swing.text.Style;
+import javax.swing.text.StyledDocument;
+
 import editor.database.attributes.CombatStat;
 import editor.database.attributes.Legality;
 import editor.database.attributes.Loyalty;
@@ -386,5 +389,30 @@ public abstract class MultiCard extends Card
     public Set<String> types()
     {
         return types.get();
+    }
+
+    @Override
+    public void formatDocument(StyledDocument document, boolean printed)
+    {
+        Style textStyle = document.getStyle("text");
+        try
+        {
+            for (int f = 0; f < faces.size(); f++)
+            {
+                formatDocument(document, printed, f);
+                if (f < faces.size() - 1)
+                    document.insertString(document.getLength(), "\n" + TEXT_SEPARATOR + "\n", textStyle);
+            }
+        }
+        catch (Exception e)
+        {
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    public void formatDocument(StyledDocument document, boolean printed, int f)
+    {
+        faces.get(f).formatDocument(document, printed);
     }
 }
